// gastromanager-frontend/src/app/services/models/user-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import {RoleResponseDto} from '../models/role-response-dto';

export interface UserResponseDto {
  email?: string;
  lastname?: string;
  name?: string;
  roles?: Array<RoleResponseDto>;
  username?: string;
  uuid?: string;
}